Refactor: Move `Self` server implementation to a dedicated module
